clean up, update pages to use same components
