+++ src/app/terms/page.tsx
@@ -1,12 +1,9 @@
 "use client";
 import Link from "next/link";
-import logoavatarise from "../../../public/logoavatarise.png";
-import Image from "next/image";
-import { useState } from "react";
+import Footer from "../components/Footer";
+import Header from "../components/Header";
 
 export default function TermsPage() {
-  const [isOpen, setIsOpen] = useState(false);
-
   return (
     <div className="min-h-screen bg-[#03090E] text-white relative overflow-hidden">
       <div>
@@ -30,13 +27,10 @@ export default function TermsPage() {
         />
       </div>
 
-      {/* Header */}
       <div className="relative z-10 pt-8 pb-12">
         <div className="max-w-4xl mx-auto px-4 sm:px-6">
           {/* Logo */}
-          <header className="relative md:absolute top-6 left-0 md:left-20 p-4 md:p-6 flex justify-center md:justify-start">
-            <Image src={logoavatarise} alt="logo" width={120} height={40} />
-          </header>
+          <Header />
 
           {/* Main Title */}
           <h1 className="text-3xl sm:text-4xl md:text-5xl font-bold text-center mb-12 mt-24">
@@ -46,7 +40,7 @@ export default function TermsPage() {
       </div>
 
       {/* Content */}
-      <div className="relative z-10 max-w-4xl mx-auto px-4 sm:px-6 pb-16">
+      <div className="relative z-10 max-w-4xl mx-auto px-4 sm:px-6 pb-16 ">
         <div className="space-y-8 text-gray-300 text-sm sm:text-base leading-relaxed">
           <div>
             <h2 className="text-lg sm:text-xl font-semibold text-white mb-2">
@@ -201,162 +195,7 @@ export default function TermsPage() {
           </Link>
         </div>
 
-        {/* Footer */}
-        <footer className="text-gray-300 mt-16">
-          <div className="pt-8" />
-          <div className="flex justify-center px-4">
-            <div
-              className="h-px bg-[#534896]"
-              style={{ width: "min(100%, 1000px)" }}
-            />
-          </div>
-
-          {/* Disclaimer Toggle */}
-          <div className="flex justify-center px-4 py-6">
-            <button
-              onClick={() => setIsOpen((o) => !o)}
-              className="flex items-center justify-center gap-2 text-sm md:text-base hover:text-white transition-colors px-4 py-2"
-            >
-              <span className="underline">Disclaimer</span>
-              <span
-                className={`text-xl leading-none transition-transform duration-300 ${
-                  isOpen ? "rotate-180" : "rotate-0"
-                }`}
-              >
-                {isOpen ? "−" : "+"}
-              </span>
-            </button>
-          </div>
-
-          <div
-            className={`overflow-hidden transition-all duration-500 ease-in-out ${
-              isOpen ? "max-h-[1200px] opacity-100" : "max-h-0 opacity-0"
-            }`}
-          >
-            <div className="max-w-6xl mx-auto px-4 pb-8 text-xs sm:text-sm leading-relaxed space-y-4">
-              <p>
-                This presentation is provided for informational and discussion
-                purposes only. It is not intended to constitute, and does not
-                constitute, an offer to sell or the solicitation of an offer to
-                buy any securities, financial instruments, or digital assets in
-                any jurisdiction. No portion of this presentation shall be
-                relied upon in connection with any investment decision.
-              </p>
-              <p>
-                Avataris AI (a brand operated by Avataris Health Innovations
-                Inc., a Canadian corporation) does not create, issue, or sell
-                the MEDX token. The MEDX token is created and issued by a
-                separate and independent entity, Medix Launch Partners, for use
-                exclusively within the Avataris Ecosystem as a utility token.
-                MEDX is not intended to function as an investment vehicle or
-                store of value and is not offered as a security or financial
-                instrument in any jurisdiction.
-              </p>
-              <p>
-                All descriptions of token utility are illustrative of current
-                and planned platform functionality only.
-              </p>
-              <p>
-                The availability of such features may vary based on region,
-                regulation, user eligibility, and ongoing platform development.
-                Use of MEDX is subject to applicable terms of service and local
-                regulations.
-              </p>
-              <p>
-                This presentation may reference digital health, AI-powered
-                diagnostics, and health-related use cases. Nothing herein should
-                be construed as medical advice.
-              </p>
-              <p>
-                All medical use cases are subject to local regulatory approval
-                and platform compliance with applicable laws, including but not
-                limited to HIPAA, PIPEDA, GDPR, and equivalent international
-                standards for data protection and privacy. Past performance,
-                product roadmaps, or adoption forecasts included in this
-                presentation are not guarantees of future results. All
-                statements herein are subject to change without notice.
-              </p>
-              <p className="font-semibold pt-2">
-                DO NOT DISTRIBUTE WITHOUT EXPRESS WRITTEN CONSENT
-                <br />
-                Confidential © Avataris Health Innovations Inc. 2025. All Rights
-                Reserved.
-              </p>
-            </div>
-          </div>
-
-          {/* Footer Layout */}
-          <div className="max-w-6xl mx-auto px-4 py-6">
-            <div className="block md:hidden space-y-4">
-              <div className="flex flex-col sm:flex-row justify-between items-start space-y-2 sm:space-y-0">
-                <div className="text-xs">info@avatarisai.com</div>
-                <div className="flex space-x-4 text-xs">
-                  <Link
-                    href="/privacy"
-                    className="hover:text-white transition-colors"
-                  >
-                    Privacy Policy
-                  </Link>
-                  <Link
-                    href="/terms"
-                    className="hover:text-white transition-colors"
-                  >
-                    Terms of Service
-                  </Link>
-                </div>
-              </div>
-
-              <div className="flex justify-between items-end">
-                <div className="flex items-center space-x-2">
-                  <Image
-                    src={logoavatarise}
-                    alt="AvatarisAI logo"
-                    width={80}
-                    height={80}
-                  />
-                </div>
-                <div className="text-right text-xs text-gray-400 space-y-1">
-                  <div>Your Health. Your Data. Your Avatar.</div>
-                  <div>© 2025 — Copyright</div>
-                </div>
-              </div>
-            </div>
-
-            <div className="hidden md:flex justify-between items-end">
-              <div className="flex flex-col space-y-2 text-sm">
-                <div>info@avatarisai.com</div>
-                <div className="flex space-x-6">
-                  <Link
-                    href="/privacy"
-                    className="hover:text-white transition-colors"
-                  >
-                    Privacy Policy
-                  </Link>
-                  <Link
-                    href="/terms"
-                    className="hover:text-white transition-colors"
-                  >
-                    Terms of Service
-                  </Link>
-                </div>
-              </div>
-              <div className="text-right">
-                <div className="flex items-center justify-end space-x-3 mb-2">
-                  <Image
-                    src={logoavatarise}
-                    alt="AvatarisAI logo"
-                    width={120}
-                    height={120}
-                  />
-                </div>
-                <div className="text-xs text-gray-400 space-y-1">
-                  <div>Your Health. Your Data. Your Avatar.</div>
-                  <div>© 2025 — Copyright</div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </footer>
+        <Footer />
       </div>
     </div>
   );
